backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.7.1
supabase==2.13.0
pdfplumber==0.11.5
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0
